requirements.txt
annotated-types==0.7.0
anyio==4.9.0
appdirs==1.4.4
argcomplete==3.6.2
asgiref==3.8.1
attr==0.3.1
attrs==25.3.0
azure-core==1.33.0
azure-storage-blob==12.25.1
black==25.1.0
bleach==5.0.1
blinker==1.9.0
boto==2.49.0
boto3==1.37.30
botocore==1.37.30
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
contourpy==1.3.1
cryptography==44.0.2
cycler==0.12.1
datamodel-code-generator==0.26.1
defusedxml==0.7.1
Deprecated==1.2.18
distro==1.9.0
Django==5.1.8
django-annoying==0.10.6
django-cors-headers==3.6.0
django-csp==3.7
django-debug-toolbar==3.2.1
django-environ==0.10.0
django-extensions==3.2.3
django-filter==24.3
django-migration-linter==5.2.0
django-model-utils==4.1.1
django-ranged-fileresponse==0.1.2
django-rq==2.10.3
django-storages==1.12.3
django-user-agents==0.4.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
dnspython==2.7.0
drf-dynamic-fields==0.3.0
drf-flex-fields==0.9.5
drf-generators==0.3.0
email_validator==2.2.0
expiringdict==1.2.2
Faker==37.1.0
filelock==3.18.0
Flask==3.1.0
fonttools==4.57.0
fsspec==2025.3.2
genson==1.3.0
google-api-core==2.24.2
google-auth==2.38.0
google-cloud-appengine-logging==1.6.1
google-cloud-audit-log==0.3.2
google-cloud-core==2.4.3
google-cloud-logging==3.11.4
google-cloud-storage==2.19.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.69.2
grpc-google-iam-v1==0.14.2
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
humansignal-drf-yasg==1.21.10.post1
idna==3.10
ijson==3.3.0
importlib_metadata==8.6.1
inflect==5.6.2
inflection==0.5.1
isodate==0.7.2
isort==5.13.2
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.9.0
jmespath==1.0.1
joblib==1.4.2
jsf==0.11.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.8
label-studio==1.17.0
label-studio-sdk==1.0.11
launchdarkly-server-sdk==8.2.1
lockfile==0.12.2
lxml==5.3.2
lxml_html_clean==0.4.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
mdurl==0.1.2
mpmath==1.3.0
mypy-extensions==1.0.0
networkx==3.4.2
nltk==3.9.1
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
openai==1.72.0
opencv-python==4.11.0.86
opentelemetry-api==1.31.1
ordered-set==4.0.2
packaging==24.2
pandas==2.2.3
pathspec==0.12.1
pillow==11.1.0
platformdirs==4.3.7
proto-plus==1.26.1
protobuf==5.29.4
psutil==7.0.0
psycopg2-binary==2.9.10
py-cpuinfo==9.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyboxen==1.3.0
pycparser==2.22
pydantic==2.11.3
pydantic_core==2.33.1
Pygments==2.19.1
PyJWT==2.10.1
pyparsing==3.2.3
pyRFC3339==2.0.1
python-dateutil==2.9.0.post0
python-json-logger==2.0.4
pytz==2022.7.1
PyYAML==6.0.2
redis==5.2.1
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-file==2.1.0
requests-mock==1.12.1
rich==14.0.0
rpds-py==0.24.0
rq==1.16.2
rsa==4.9
rstr==3.2.2
rules==3.4
s3transfer==0.11.4
scipy==1.15.2
seaborn==0.13.2
semver==3.0.4
sentry-sdk==2.25.1
setuptools==78.1.0
six==1.17.0
smart-open==7.1.0
sniffio==1.3.1
sqlparse==0.5.3
sympy==1.13.1
tldextract==5.2.0
toml==0.10.2
torch==2.6.0
torchvision==0.21.0
tqdm==4.67.1
triton==3.2.0
typing-inspection==0.4.0
typing_extensions==4.13.1
tzdata==2025.2
ua-parser==1.0.1
ua-parser-builtins==0.18.0.post1
ujson==5.10.0
ultralytics==8.3.108
ultralytics-thop==2.0.14
uritemplate==4.1.1
urllib3==1.26.20
user-agents==2.2.0
webencodings==0.5.1
Werkzeug==3.1.3
wheel==0.40.0
wrapt==1.17.2
xmljson==0.2.1
zipp==3.21.0
